% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_types.R
\name{auto_var_types}
\alias{auto_var_types}
\title{Automaticall assign var_types to the attributes}
\usage{
auto_var_types(tbl)
}
\arguments{
\item{tbl}{The imported data.frame}
}
\description{
If the user did not provide var_types, the function will
preassume every variables to be quantitative variables. If a variable's type
is character but either the user or the automatic step said it's a
quantitative variable, the var_types attribute for that variable will be
overwritten as categorical variable. At the same time, a warning message
will be printed on the screen.
}

