% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ez_summarise_module_categorical.r
\name{ez_summarise_categorical}
\alias{ez_summarise_categorical}
\title{Easy summary for categorical data}
\usage{
ez_summarise_categorical(tbl, n = FALSE, count = TRUE, p = TRUE,
  P = FALSE, round.N = 3)
}
\arguments{
\item{tbl}{The input matrix of data you would like to analyze.}

\item{n}{n is a True/False switch that controls whether total counts(N) should be included in
the output.}

\item{count}{a T/F switch to control if counts should be included}

\item{p}{a T/F switch to control if proportion should be included}

\item{P}{a T/F switch to control if Percentage should be included. This will be a
character output}

\item{round.N}{Rounding number.}
}
\value{
This function will organize all the results into one dataframe. If there are
any group_by variables, the first few columns will be them. After these, the varible
x will be the one listing all the categorical options in a format like "variable_option".
The stats summaries are listed in the last few columns.
}
\description{
This function provided an easy-to-use way to display simple statistical
summary for categorical data. It can be used together with \code{\link[dplyr]{select}}
and \code{\link[dplyr]{group_by}}. If the piece of data passed into the function has
one/multiple group_by variables, the percentage and count will be calculated within
the defined groups.
}
\examples{
mtcars \%>\% group_by(am) \%>\% select(cyl, gear, carb) \%>\% ez_summarise_categorical()
mtcars \%>\% select(cyl, gear, carb) \%>\% ez_summarise_categorical(n=TRUE, round.N = 2)
}

