% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ezsummary.r
\name{ezsummary}
\alias{ezsummary}
\title{Quick and Easy summarise function}
\usage{
ezsummary(tbl, n = F, round.N = 3)
}
\arguments{
\item{tbl}{The input matrix of data you would like to analyze.}

\item{n}{n is a True/False switch that controls whether counts(N) should be included in
the output}

\item{round.N}{Rounding Number}
}
\description{
Quick and Easy summarise function
}

