% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simple_summary_module_categorical.r
\name{simple_summary_categorical}
\alias{simple_summary_categorical}
\title{Simple summary for categorical data}
\usage{
simple_summary_categorical(tbl, n = F, round.N = 3)
}
\arguments{
\item{tbl}{The input matrix of data you would like to analyze.}

\item{n}{n is a True/False switch that controls whether total counts(N) should be included in
the output.}

\item{round.N}{Rounding number.}
}
\value{
This function will organize all the results into one dataframe. If there are
any group_by variables, the first few columns will be them. After these, the varible
x will be the one listing all the categorical options in a format like "variable_option".
The stats summaries are listed in the last few columns.
}
\description{
This function provided an easy-to-use way to display simple statistical
summary for categorical data. It can be used together with \code{\link{dplyr::select}}
and \code{\link{dplyr::group_by}}. If the piece of data passed into the function has
one/multiple group_by variables, the percentage and count will be calculated within
the defined groups.
}
\examples{
mtcars \%>\% group_by(am) \%>\% select(cyl, gear, carb) \%>\% simple_summary_categorical()
mtcars \%>\% select(cyl, gear, carb) \%>\% simple_summary_categorical(n=T, round.N = 2)
}

