% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simple_summary_module_binary.r, R/simple_summary_module_quantitative.r
\name{simple_summary_binary}
\alias{simple_summary_binary}
\alias{simple_summary_quantitative}
\title{Simple Summary for Quantitative/binary variables}
\usage{
simple_summary_binary(tbl, n = F)

simple_summary_quantitative(tbl, n = F, round.N = 3)
}
\arguments{
\item{tbl}{The input matrix of data you would like to analyze.}

\item{n}{n is a True/False switch that controls whether counts(N) should be included in
the output}

\item{round.N}{Rounding Number}
}
\value{
It will return in the same format as a summarise_each function does
}
\description{
Function simple_summary_quantitative provides simple summary (Mean and
standard deviation with/without N) for quantitative data while function simple_summary_binary
provides simple summary (freq and percentage with/without total counts) for binary data.
These two function are simply wrappers outside of a summarise_each function. If
we just want to know the most basic statistical summary, this function can save us some
typing time. It also provide the option to include number of subjects inside the analyses.
}
\examples{
mtcars \%>\% group_by(am) \%>\% select(mpg, wt, qsec) \%>\% simple_summary_quantitative()
}

