% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simple_summary_module_categorical.r
\name{count_percentage}
\alias{count_percentage}
\alias{count_percentage_}
\title{Counts and Percentages}
\usage{
count_percentage(data_vector, n = F, round.N = 3)

count_percentage_(a, n = F, round.N = 3)
}
\description{
This is a calculating module used by \code{\link{simple_summary_categorical}} but
it can also be used independently.
}

