% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezsummary.r
\name{ezsummary}
\alias{ezsummary}
\title{Quick and Easy summarise function}
\usage{
ezsummary(tbl, ...)
}
\arguments{
\item{tbl}{A vector, a data.frame or a \code{dplyr} \code{tbl}.}

\item{...}{Arguments that can be passed to \code{ezsummary_q()} and
\code{ezsummary_c()}}
}
\description{
Quick and Easy summarise function
}

