% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezsummary_quantitative.r
\name{ezsummary_quantitative}
\alias{ezsummary_q}
\alias{ezsummary_quantitative}
\title{Easily summarize quantitative data}
\usage{
ezsummary_quantitative(tbl, total = FALSE, n = FALSE, missing = FALSE,
  mean = TRUE, sd = TRUE, sem = FALSE, median = FALSE,
  quantile = FALSE, extra = NULL, digits = 3, rounding_type = c("round",
  "signif", "ceiling", "floor"), round.N = 3, flavor = c("long", "wide"),
  fill = 0, unit_markup = NULL)

ezsummary_q(tbl, total = FALSE, n = FALSE, missing = FALSE, mean = TRUE,
  sd = TRUE, sem = FALSE, median = FALSE, quantile = FALSE,
  extra = NULL, digits = 3, rounding_type = c("round", "signif",
  "ceiling", "floor"), round.N = 3, flavor = c("long", "wide"), fill = 0,
  unit_markup = NULL)
}
\arguments{
\item{tbl}{A vector, a data.frame or a \code{dplyr} \code{tbl}.}

\item{total}{a T/F value; total counts of records including both missing
and read data records. Default is \code{FALSE}.}

\item{n}{A T/F value; total counts of records that is not missing. Default
is \code{FALSE}.}

\item{missing}{a T/F value; total counts of records that went missing(
\code{NA}). Default is \code{FALSE}.}

\item{mean}{A T/F value; the average of a set of data. Default value is
\code{TRUE}.}

\item{sd}{A T/F value; the standard deviation of a set of data. Default value
is \code{TRUE}.}

\item{sem}{A T/F value; the standard error of the mean of a set of data.
Default value is \code{FALSE}.}

\item{median}{A T/F value; the median of a set of data. Default value is
\code{FALSE}.}

\item{quantile}{A T/F value controlling 5 outputs; the 0\%, 25\%, 50\%, 75\%
and 100\% percentile of a set of data. Default value is \code{FALSE}.}

\item{extra}{A character vector offering extra customizability to this
function. Please see Details for detail.}

\item{digits}{A numeric value determining the rounding digits; Replacement
for \code{round.N}. Default setting is to read from \code{getOption()}.}

\item{rounding_type}{A character string determining the rounding method;
possible values are \code{round}, \code{signif}, \code{ceiling} and
\code{floor}. When \code{ceiling} or \code{floor} is selected, \code{digits}
won't have any effect.}

\item{round.N}{Deprecated; Will change the value of \code{rounding_type} if
used in this version.}

\item{flavor}{A character string with two possible inputs: "long" and "wide".
"Long" is the default setting which will put grouping information on the left
side of the table. It is more machine readable and is good to be passed into
the next analytical stage if needed. "Wide" is more print ready (except for
column names, which you can fix in the next step, or fix in LaTex or
packages like \code{htmlTable}). In the "wide" mode, the analyzed variable
will be the only "ID" variable and all the stats values will be presented
ogranized by the grouping variables (if any). If there is no grouping, the
outputs of "wide" and "long" will be the same.}

\item{fill}{If set, missing values created by the "wide" flavor will be
replaced with this value. Please check \code{\link[tidyr]{spread}} for
details. Default value is \code{0}}

\item{unit_markup}{When unit_markup is not NULL, it will call the ezmarkup
function and perform column combination here. To make everyone's life
easier, I'm using the term "unit" here. Each unit mean each group of
statistical summary results. If you want to know mean and stand deviation,
these two values are your units so you can put something like "[. (.)]" there
#' @param P Deprecated; Will change the value of \code{p_type} if used in this
version.}
}
\description{
\code{ezsummary_quantitative()} summarizes quantitative data.
}
\examples{
library(dplyr)
mtcars \%>\%
  group_by(am) \%>\%
  select(mpg, wt, qsec) \%>\%
  ezsummary_quantitative()

}

